Fixed time costants to be long, not float

+++ 82-tdd-deathnote/src/main/java/it/unibo/deathnote/impl/DeathNoteImpl.java
@@ -10,8 +10,8 @@ import it.unibo.deathnote.api.DeathNote;
  */
 public final class DeathNoteImpl implements DeathNote {
     private static final int RULE_OFFSET = 1;
-    private static float DEATH_MILLIS_TIME = 40;
-    private static float DETAILS_MILLIS_TIME = 6000 + DEATH_MILLIS_TIME;
+    private static long DEATH_MILLIS_TIME = 40;
+    private static long DETAILS_MILLIS_TIME = 6000 + DEATH_MILLIS_TIME;
 
     private Map<String, Death> deathNote;
     private String lastNameInserted;
